% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{combine_boot}
\alias{combine_boot}
\title{combine_boot: combine imputed bootstrap datasets}
\usage{
combine_boot(
  ls_df,
  col_con,
  col_dis = c(),
  col_cat = c(),
  num_row_origin,
  method = "onehot",
  dict_cat = NULL,
  var_cat = "unalike"
)
}
\arguments{
\item{ls_df}{A list of imputed bootstrapped dataframes.}

\item{col_con}{Continous columns index.}

\item{col_dis}{Discret columns index.}

\item{col_cat}{Categorical columns index.}

\item{num_row_origin}{Number of rows in the original incomplete dataframe before bootstrapping.}

\item{method}{The encoded method of categorical columns in the imputed dataframes.
This function is coded for both "onehot" and "factor" encoded situations.
When \code{method} = 'onehot', \code{combine_boot} averages the probability vectors over the $B$ imputed datasets for the same observation,
then choose the position of maximum probability as the predicted category in the final result.
 When \code{method} = 'factor', for each observation, \code{combine_boot}
choose the mode value over the imputed dataframes as the predicted category.}

\item{dict_cat}{The dictionary of categorical columns names if "onehot" method is applied.
For example, it could be list("Y7"=c("Y7_1","Y7_2"), "Y8"=c("Y8_1","Y8_2","Y8_3")).}

\item{var_cat}{The method of variance calculation for the categorical columns.
"unalike" will lead to the calculation of unalikeability, while "wilcox_va" will lead to the calculation of Wilcox index: VarNC.}
}
\value{
\code{df_result_disj} The final imputed dataframe with the categorical columns in onehot form.

\code{df_result_var_disj} The variance matrix for the final imputation dataframe with the categorical columns in onehot form.

\code{df_result} The final imputed dataframe with the categorical columns in factor form.

\code{df_result_var} The variance matrix for the final imputation dataframe with the categorical columns in factor form.
}
\description{
\code{combine_boot} function combines several imputed bootstrapped dataframes
into the final imputed dataframe and provide the variance for each imputed value.
}
\references{
Statistical Analysis with Missing Data, by Little and Rubin, 2002
}
