% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{combine_jack}
\alias{combine_jack}
\title{combine_jack: combine imputed jackknife datasets}
\usage{
combine_jack(
  ls_df,
  df_full,
  col_con,
  col_dis = c(),
  col_cat = c(),
  method = "onehot",
  dict_cat = NULL,
  var_cat = "unalike"
)
}
\arguments{
\item{ls_df}{A list of imputed jackknifed dataframes.}

\item{df_full}{Imputation result on the original dataframe before jackknife resampling.}

\item{col_con}{Continous columns index.}

\item{col_dis}{Discret columns index.}

\item{col_cat}{Categorical columns index.}

\item{method}{The encoded method of categorical columns in the imputed dataframes. 
This function is only coded for "onehot" situation.}

\item{dict_cat}{The dictionary of categorical columns names if "onehot" method is applied.
For example, it could be list("Y7"=c("Y7_1","Y7_2"), "Y8"=c("Y8_1","Y8_2","Y8_3")).}

\item{var_cat}{The method of variance calculation for the categorical columns. 
"unalike" will lead to the calculation of unalikeability, while "wilcox_va" will lead to the calculation of Wilcox index: VarNC.}

\item{num_row_origin}{Number of rows in the original incomplete dataframe before bootstrapping.}
}
\value{
\code{df_result_disj} The final imputed dataframe with the categorical columns in onehot form.

\code{df_result_var_disj} The variance matrix for the final imputation dataframe with the categorical columns in onehot form.

\code{df_result} The final imputed dataframe with the categorical columns in factor form.

\code{df_result_var} The variance matrix for the final imputation dataframe with the categorical columns in factor form.
}
\description{
\code{combine_jack} function combines several imputed jackknifed dataframes 
into the final imputed dataframe and provide the variance for each imputed value.
}
\references{
Statistical Analysis with Missing Data, by Little and Rubin, 2002
}
