% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{dens_comp}
\alias{dens_comp}
\title{dens_comp}
\usage{
dens_comp(df_comp, df_imp)
}
\arguments{
\item{df_comp}{One dataset.}

\item{df_imp}{Another dataset with the same columns as \code{df_comp}.}
}
\value{
The comparison of the density distribution for each column.
}
\description{
There are density comparing functions for some of the imputation method, 
but here we try to define the function that could be used for every method.
}
