% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dict_level}
\alias{dict_level}
\title{dict_level}
\usage{
dict_level(df, col_cat)
}
\arguments{
\item{df}{Dataframe.}

\item{col_cat}{categorical column index}
}
\value{
A dictionary that connects categorical columns with their original levels.
}
\description{
Create the dictionary for categorical columns levels. For example, {Y7:A,B,C,...}.
}
