% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dict_onehot}
\alias{dict_onehot}
\title{dict_onehot}
\usage{
dict_onehot(df, col_cat)
}
\arguments{
\item{df}{Dataframe.}

\item{col_cat}{categorical column index}
}
\value{
A dictionary that connects onehot form and factor form of the categorical columns.
}
\description{
Create the dictionary for categorical columns. For example, {Y7:Y7_1,Y7_2,...}.
}
