% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mod.R
\name{em_mod}
\alias{em_mod}
\title{em_mod: modified EM Imputation with probability vector}
\usage{
em_mod(df, col_cat)
}
\arguments{
\item{df}{Data matrix with missing values.}

\item{col_cat}{Categorical columns index}
}
\value{
\code{ximp} imputed data matrix.

\code{ximp.disj} imputed data matrix of same type as 'ximp' for the numeric columns.
 For the categorical columns, the prediction of probability for each category is shown in form of onehot probability vector.
}
\description{
\code{em_mod} is a em imputation function that returns categorical columns results both in factor and in onehot probability vector form.
Please find the detailed documentation of \code{em.mix} and \code{imp.mix} in the 'mix' package. Only the modifications are explained on this page.
After the estimation of parameter \code{pi} in \code{em.mix}, we change it into a tensor (multidimensional array) and
extract the probability vector from this tensor with the help of function \code{prob_vector_cat}.
}
