% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missMDA_function.R
\name{estim_ncpFAMD_mod}
\alias{estim_ncpFAMD_mod}
\title{estim_ncpFAMD_mod}
\usage{
estim_ncpFAMD_mod(
  don,
  ncp.min = 0,
  ncp.max = 5,
  method = c("Regularized", "EM"),
  method.cv = c("Kfold", "loo"),
  nbsim = 100,
  pNA = 0.05,
  ind.sup = NULL,
  sup.var = NULL,
  threshold = 1e-04,
  verbose = TRUE,
  maxiter = 1000
)
}
\arguments{
\item{don}{a data.frame with categorical variables; with missing entries or not.}

\item{ncp.min}{integer corresponding to the minimum number of components to test.}

\item{ncp.max}{integer corresponding to the maximum number of components to test.}

\item{method}{"Regularized" by default or "EM".}

\item{method.cv}{"Kfold" for cross-validation or "loo" for leave-one-out.}

\item{nbsim}{number of simulations, useful only if method.cv="Kfold".}

\item{pNA}{percentage of missing values added in the data set, useful only if method.cv="Kfold.}

\item{ind.sup}{a vector indicating the indexes of the supplementary individuals.}

\item{sup.var}{a vector indicating the indexes of the supplementary variables (quantitative and categorical).}

\item{threshold}{the threshold for assessing convergence}

\item{verbose}{boolean. TRUE means that a progressbar is writtent.}

\item{maxiter}{max iteration number for \code{imputeFAMD_mod}}
}
\value{
\code{ncp} 	the number of components retained for the FAMD.

\code{criterion} the criterion (the MSEP) calculated for each number of components.
}
\description{
Estimate the number of dimensions for the Factorial Analysis of Mixed Data by cross-validation. 
This function is nearly identical with \code{estim_ncpFAMD} function in 'missMDA' package. The only difference is 
that in this function \code{imputeFAMD_mod} is used, which then calls \code{imputeMFA_mod}, then \code{impute_mod}.
 In \code{impute_mod}, some changes have been made to avoid the convergence error.
}
\references{
Audigier, V., Husson, F. & Josse, J. (2014). A principal components method to impute mixed data. Advances in Data Analysis and Classification.
}
