% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_encode}
\alias{factor_encode}
\title{factor_encode}
\usage{
factor_encode(df, idx_col_cat)
}
\arguments{
\item{df}{Dataframe}

\item{idx_col_cat}{Categorical columns index}
}
\value{
Dataframe after factor encoding the categorical columns
}
\description{
Factor encode the categorical columns in a dataframe
}
