% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_ordinal_encode}
\alias{factor_ordinal_encode}
\title{factor_ordinal_encode}
\usage{
factor_ordinal_encode(df, idx_col_cat)
}
\arguments{
\item{df}{Dataframe}

\item{idx_col_cat}{Categorical columns index}
}
\value{
Dataframe after encoding the categorical columns
}
\description{
Change the levels of categorical columns into integer numbers
}
