% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_missingness.R
\name{generate_miss}
\alias{generate_miss}
\title{generate_miss: Generate missing values with different mechanisms}
\usage{
generate_miss(df, miss_perc, mechanism = "MCAR", mar2.col.ctrl = 1)
}
\arguments{
\item{df}{Complete dataframe.}

\item{miss_perc}{Desired percentage of missing values. Note that all the mechanisms could only 
approach this proportion. The real proportion of missingness will be returned with the incomplete dataframe.
When the number of columns is small, some mechanisms could be incompatible with large missing percentage.}

\item{mechanism}{Desired missing mechanism.
\itemize{
 \item \strong{MCAR} generates missing values by missing completely at random mechanism with Bernouilli distribution. 
 \item \strong{MAR1} generates missing values by missing at random mechanism with logistic regression on the observed data.
 \item \strong{MAR2} generates missing values by missing at random mechanism with censoring algorithm. 
The missingness in every other column depends on the quantile of one specified complete column \code{mar2.col.ctrl}.
For example, on row i, Y2[i] will be removed if Y1[i]<q(30\%) of Y1.
 \item \strong{MAR3} generates missing values by missing at random mechanism with monotone censoring mechanism.
The missingness in every column depends on the quantile of the observed data of the column before.
For example, on row i, Y2[i],Y3[i],...,Yn[i] will be removed if Y1[i]<q(30\%) of Y1.
And then on row s, Y3[s],...,Yn[s] will be removed if Y2[s]<q(30\%) of observed Y2 (Those who are not removed in step 1).
 \item \strong{MNAR1} generates missing values by missing not at random mechanism with logistic regression on the observed and missing part of data.
 \item \strong{MNAR2} generates missing values by missing not at random mechanism with censoring mechanism.
 For example, for each column j, on row i, Yj[i] will be removed if Yj[i]<q(30\%) of Y1.
}}

\item{mar2.col.ctrl}{Control column in mechanism MAR2}
}
\value{
\code{X.incomp} Generated incomplete dataframe.

\code{R.mask} Mask of missingness. R[2,3]=1 means that df[2,3] is missing.

\code{real_miss_perc} Real proportion of missingness.
}
\description{
\code{generate_miss} function generates missing values in a complete 
dataframe with the chosen mechanism and the chosen proportion of missingness.
}
\examples{
n = 10000
mu.X = c(1, 2, 3)
Sigma.X = matrix(c(9, 3, 2, 3, 4, 0, 2, 0, 1), nrow = 3)
X.complete.cont = MASS::mvrnorm(n, mu.X, Sigma.X)
rs = generate_miss(X.complete.cont, 0.5, mechanism = "MNAR2")
rs$X.incomp
}
\references{
\itemize{
\item \url{https://rmisstastic.netlify.app/how-to/generate/misssimul}
\item \url{https://cran.r-project.org/web/packages/missMethods/vignettes/Generating-missing-values.html}
\item Santos, M. S., R. C. Pereira, A. F. Costa, J. P. Soares, J. Santos, and P. H. Abreu. 2019. Generating Synthetic Missing Data: A Review by Missing Mechanism. IEEE Access 7: 11651–67. \url{https://doi.org/10.1109/ACCESS.2019.2891360}.
}
}
