% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_missingness.R
\name{generate_miss_ls}
\alias{generate_miss_ls}
\title{generate_miss_ls: Generate a list of incomplete dataframes with different missing mechanisms}
\usage{
generate_miss_ls(df, miss_perc)
}
\arguments{
\item{df}{Complete dataframe.}

\item{miss_perc}{Desired percentage of missing values. Note that all the mechanisms could only
approach this proportion. The real proportion of missingness will be returned with the incomplete dataframe.
When the number of columns is small, some mechanisms could be incompatible with large missing percentage.}
}
\value{
\code{mcar} Incomplete dataframe object with MCAR mechanism. The detailed description could be
found in \code{generate_miss} documentation. Each object contains three attributes:
\code{X.incomp}, \code{R.mask}, and \code{real_miss_perc}

\code{mar1} Incomplete dataframe object with MAR1 mechanism.

\code{mar2} Incomplete dataframe object with MAR2 mechanism.

\code{mar3} Incomplete dataframe object with MAR3 mechanism.

\code{mnar1} Incomplete dataframe object with MNAR1 mechanism.

\code{mnar2} Incomplete dataframe object with MNAR2 mechanism.
}
\description{
\code{generate_miss_ls} function generates a list of incomplete dataframes. Missing values are generated in the given complete
dataframe with the all mechanisms in \code{generate_miss} function and the chosen proportion of missingness.
}
