% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missMDA_function.R
\name{imputeFAMD_mod}
\alias{imputeFAMD_mod}
\title{imputeFAMD_mod}
\usage{
imputeFAMD_mod(
  X,
  ncp = 2,
  method = c("Regularized", "EM"),
  row.w = NULL,
  coeff.ridge = 1,
  threshold = 1e-06,
  ind.sup = NULL,
  sup.var = NULL,
  seed = NULL,
  maxiter = 1000,
  ...
)
}
\arguments{
\item{X}{a data.frame with continuous and categorical variables containing missing values.}

\item{ncp}{integer corresponding to the number of components used to predict the missing entries.}

\item{method}{"Regularized" by default or "EM".}

\item{row.w}{row weights (by default, uniform row weights).}

\item{coeff.ridge}{1 by default to perform the regularized imputeFAMD algorithm; useful only if method="Regularized". 
Other regularization terms can be implemented by setting the value to less than 1 in order to regularized less 
(to get closer to the results of the EM method) or more than 1 to regularized more.}

\item{threshold}{the threshold for assessing convergence}

\item{ind.sup}{a vector indicating the indexes of the supplementary individuals.}

\item{sup.var}{a vector indicating the indexes of the supplementary variables (quantitative and categorical).}

\item{seed}{integer, by default seed = NULL implies that missing values are initially imputed by the 
mean of each variable for the continuous variables and by the proportion of the category for the categorical variables 
coded with indicator matrices of dummy variables. Other values leads to a random initialization.}

\item{maxiter}{max iteration number for \code{imputeFAMD_mod}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{tab.disj} 	the imputed matrix; the observed values are kept for the non-missing entries and 
the missing values are replaced by the predicted ones. The categorical variables are coded with the indicator matrix of dummy variables. 
In this indicator matrix, the imputed values are real numbers but they met the constraint that the 
sum of the entries corresponding to one individual and one variable is equal to one. Consequently they can be seen as 
degree of membership to the corresponding category.

\code{completeObs} the mixed imputed dataset; the observed values are kept for the non-missing entries and the missing values are replaced by the predicted ones. 
For the continuous variables, the values are the same as in the tab.disj output; for the categorical variables missing values are imputed with the 
most plausible categories according to the values in the tab.disj output.

\code{call} the matched call.
}
\description{
Impute the missing values of a mixed dataset (with continuous and categorical variables) using the principal 
component method "factorial analysis for mixed data" (FAMD). 
Can be used as a preliminary step before performing FAMD on an incomplete dataset.
 
This function is nearly identical with \code{imputeFAMD} function in 'missMDA' package. The only difference is 
that in this function \code{imputeMFA_mod}is used, which then calls \code{impute_mod}.
 In \code{impute_mod}, some changes have been made to avoid the convergence error.
}
\references{
Audigier, V., Husson, F. & Josse, J. (2013). A principal components method to impute mixed data. Advances in Data Analysis and Classification, 10(1), 5-26. 
\url{https://arxiv.org/abs/1301.4797#'}
}
