% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{ls_F1}
\alias{ls_F1}
\title{List of MSE}
\usage{
ls_F1(
  df_comp,
  ls_df_imp,
  mask,
  col_cat,
  resample_method = "bootstrap",
  combine_method = "onehot",
  dict_cat = NULL
)
}
\arguments{
\item{df_comp}{The original complete dataset.}

\item{ls_df_imp}{List of imputed dataset.}

\item{mask}{Mask of missingness (1 means missing value and 0 means observed value).}

\item{col_cat}{Indices of categorical columns.}

\item{resample_method}{Default value is 'bootstrap', could also be 'jackknife' or 'none'.}

\item{combine_method}{When \code{resample_method} = 'bootstrap', \code{combine_method} could be 'factor' or 'onehot'.
When \code{method} = 'onehot', \code{ls_F1} takes the average of the onehot probability vector for each observation,
then choose the position of maximum probability as the predicted category. When \code{method} = 'factor', for each observation, \code{ls_F1}
choose the mode value over the imputed dataframes as the predicted category.}

\item{dict_cat}{The dictionary of categorical columns names if "onehot" method is applied.
For example, it could be list("Y7"=c("Y7_1","Y7_2"), "Y8"=c("Y8_1","Y8_2","Y8_3")).}
}
\value{
\code{list_F1} List of F1 corresponding to the given list of imputed datasets.

\code{Mean_F1} Mean value of F1.

\code{Variance_F1} Variance of F1.
}
\description{
\code{ls_F1} is a function that returns a list of F1-Score corresponding to the given list of imputed datasets.
\code{resample_method} is needed because with 'bootstrap' method, we could have repeated lines in the imputed datasets,
and with both 'jackknife' and 'bootstrap', the imputed datasets could not cover all the lines.
}
