% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_em_amelia.R
\name{MI_EM_amelia}
\alias{MI_EM_amelia}
\title{MI_EM_amelia: Multiple imputation with EM}
\usage{
MI_EM_amelia(df_with_mv, col_num, col_cat = NULL, num_imp = 5)
}
\arguments{
\item{df_with_mv}{Data matrix with missing values.}

\item{col_num}{Numerical columns index}

\item{col_cat}{Categorical columns index}

\item{num_imp}{Number of multiple imputations}
}
\value{
\code{ximp} Final imputed data matrix.

\code{ximp.disj} Final imputed data matrix of same type as 'ximp' for the numeric columns.
 For the categorical columns, the prediction of probability for each category is shown in form of onehot probability vector.

\code{ls_ximp} List imputed data matrix from multiple imputation procedure.

\code{ls_ximp.disj} List imputed data matrix from multiple imputation procedure, with categorical columns
in onehot probability vector form.
}
\description{
\code{MI_EM_amelia} is a multiple imputation function with EM algorithm.
This function returns both multiple imputation results of the original incomplete data and
the final imputed data (derived from multiple imputation results by Robin's Rule).
As for categorical columns, both factor form and onehot probability vector form are returned.

More details about the EM implementation could be found in the documentation of function \code{amelia} from
 'Amelia' package.
}
\references{
Honaker, J., King, G., Blackwell, M. (2011).
Amelia II: A Program for Missing Data.
\emph{Journal of Statistical Software}, \bold{45(7)}, 1--47.
\url{http://www.jstatsoft.org/v45/i07/}.

Statistical Analysis with Missing Data, by Little and Rubin, 2002
}
