% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_ranger.R
\name{MI_missRanger}
\alias{MI_missRanger}
\title{MI_missRanger}
\usage{
MI_missRanger(
  data,
  formula = . ~ .,
  pmm.k = 0L,
  maxiter = 10L,
  seed = NULL,
  verbose = 1,
  returnOOB = FALSE,
  case.weights = NULL,
  col_cat = c(),
  num_mi = 5,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} with missing values to impute.}

\item{formula}{A two-sided formula specifying variables to be imputed (left hand side) and variables used to impute (right hand side). Defaults to . ~ ., i.e. use all variables to impute all variables.
If e.g. all variables (with missings) should be imputed by all variables except variable "ID", use . ~ . - ID. Note that a "." is evaluated separately for each side of the formula. Further note that variables
with missings must appear in the left hand side if they should be used on the right hand side.}

\item{pmm.k}{Number of candidate non-missing values to sample from in the predictive mean matching steps. 0 to avoid this step.}

\item{maxiter}{Maximum number of chaining iterations.}

\item{seed}{Integer seed to initialize the random generator.}

\item{verbose}{Controls how much info is printed to screen. 0 to print nothing. 1 (default) to print a "." per iteration and variable, 2 to print the OOB prediction error per iteration and variable (1 minus R-squared for regression).
Furthermore, if \code{verbose} is positive, the variables used for imputation are listed as well as the variables to be imputed (in the imputation order). This will be useful to detect if some variables are unexpectedly skipped.}

\item{returnOOB}{Logical flag. If TRUE, the final average out-of-bag prediction error is added to the output as attribute "oob". This does not work in the special case when the variables are imputed univariately.}

\item{case.weights}{Vector with non-negative case weights.}

\item{col_cat}{Indices of categorical columns}

\item{num_mi}{Number of multiple imputation}

\item{...}{Arguments passed to \code{ranger()}. If the data set is large, better use less trees (e.g. \code{num.trees = 20}) and/or a low value of \code{sample.fraction}.
The following arguments are e.g. incompatible with \code{ranger}: \code{write.forest}, \code{probability}, \code{split.select.weights}, \code{dependent.variable.name}, and \code{classification}.}
}
\value{
\code{ximp} Final imputed dataset.

\code{ximp.disj} Final disjunctive imputed dataset.

\code{ls_imputations} List of imputed dataset from multiple imputation.

\code{ls_imputations.disj} List of disjunctive imputed dataset from multiple imputation.
}
\description{
\code{MI_missRanger} is a function of multiple imputation with \code{missRanger} method.

 In \code{missRanger_mod_draw}, for a certain prediction,
 instead of taking average of the prediction result from each tree of the random forest, we draw one result from the empirical 
 distribution constructed by predictions of trees. The other steps of the imputation are identical as those of \code{missRanger} 
 from 'missRanger' package.
 
 \code{MI_missRanger} takes all the imputation results from \code{missRanger_mod_draw} and combine them with Rubin's Rule 
 to generate the final imputed data set.
}
