% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_forest_function.R
\name{missForest_mod}
\alias{missForest_mod}
\title{missForest_mod: modified missForest with onehot probability}
\usage{
missForest_mod(
  xmis,
  maxiter = 10,
  ntree = 100,
  variablewise = FALSE,
  decreasing = FALSE,
  verbose = FALSE,
  mtry = floor(sqrt(ncol(xmis))),
  replace = TRUE,
  classwt = NULL,
  cutoff = NULL,
  strata = NULL,
  sampsize = NULL,
  nodesize = NULL,
  maxnodes = NULL,
  xtrue = NA,
  parallelize = c("no", "variables", "forests"),
  col_cat = c()
)
}
\arguments{
\item{xmis}{data matrix with missing values.}

\item{maxiter}{stop after how many iterations (default = 10).}

\item{ntree}{how many trees are grown in the forest (default = 100).}

\item{variablewise}{(boolean) return OOB errors for each variable separately.}

\item{decreasing}{(boolean) if TRUE the columns are sorted with decreasing amount of missing values.}

\item{verbose}{(boolean) if TRUE then missForest returns error estimates, runtime and if available true error during iterations.}

\item{mtry}{how many variables should be tried randomly at each node.}

\item{replace}{(boolean) if TRUE bootstrap sampling (with replacements) is performed, else subsampling (without replacements).}

\item{classwt}{list of priors of the classes in the categorical variables.}

\item{cutoff}{list of class cutoffs for each categorical variable.}

\item{strata}{list of (factor) variables used for stratified sampling.}

\item{sampsize}{list of size(s) of sample to draw}

\item{nodesize}{minimum size of terminal nodes, vector of length 2, with
number for continuous variables in the first entry and
number for categorical variables in the second entry.}

\item{maxnodes}{maximum number of terminal nodes for individual trees}

\item{xtrue}{complete data matrix}

\item{parallelize}{TODO}

\item{col_cat}{index of categorical columns.}
}
\value{
\code{ximp} imputed data matrix of same type as 'xmis'.

\code{ximp.disj} imputed data matrix of same type as 'xmis' for the numeric columns.
 For the categorical columns, the prediction of probability for each category is shown in form of onehot vector.

\code{OOBerror} estimated OOB imputation error. For the set of continuous variables in 'xmis' the NRMSE and
for the set of categorical variables the proportion of falsely classified entries is returned.
See Details for the exact definition of these error measures.
If 'variablewise' is set to 'TRUE' then this will be a vector of length 'p' where 'p' is the number of variables and the entries will be the OOB error for each variable separately.

\code{error} true imputation error. This is only available if 'xtrue' was supplied.
The error measures are the same as for 'OOBerror'.
}
\description{
\code{missForest_mod} is a modified version of the function \code{missForest} by Daniel Stekhoven.
Please find the detailed documentation of \code{missForest} in the missForest package. Only the modifications are explained on this page.
The original \code{missForest} function returns the final imputation result after convergence or \code{maxiter} iterations.
The results of categorical columns are returned in form of vector. In \code{missForest_mod} function, during the last iteration,
not only the final result, but also the onehot probability for each category is returned.
}
