% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode_cat}
\alias{Mode_cat}
\title{Mode_cat}
\usage{
Mode_cat(x)
}
\arguments{
\item{x}{Vector of categorical results. For example c('A','B','A').}
}
\value{
Most frequent category in vector \code{x}.
}
\description{
Find the most frequent result for one categorical variable.
This function is used in combine_boot with method = 'factor'
}
