% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_missingness.R
\name{monot_quantil}
\alias{monot_quantil}
\title{monot_quantil}
\usage{
monot_quantil(miss_perc, num_col)
}
\arguments{
\item{miss_perc}{Desired proportion of missingness}

\item{num_col}{Number of columns}
}
\value{
The required percentage for the quantil in MAR3. Solution to the function (1-x)^p + (1-m)*p*x -1 = 0 in (0, 1)
}
\description{
Calculate the missing proportion in MAR3, used in 'generate_miss'.
Solve the function (1-x)^p + (1-m)*p*x -1 = 0 in (0, 1), where m = miss_perc, p = num_col
}
