% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_num}
\alias{normalize_num}
\title{normalize_num}
\usage{
normalize_num(df, idx_col_num)
}
\arguments{
\item{df}{Dataframe}

\item{idx_col_num}{Numerical columns index}
}
\value{
Dataframe after scaling the numerical columns
}
\description{
Scale the numerical columns in a dataframe
}
