% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ordinal_encode}
\alias{ordinal_encode}
\title{ordinal_encode}
\usage{
ordinal_encode(df, idx_col_cat)
}
\arguments{
\item{df}{Dataframe}

\item{idx_col_cat}{Categorical columns index}
}
\value{
Dataframe after ordinal encoding the categorical columns
}
\description{
Ordinal encode the categorical columns in a dataframe
}
