% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mod.R
\name{prepare_df_for_em}
\alias{prepare_df_for_em}
\title{prepare_df_for_em}
\usage{
prepare_df_for_em(df, col_cat)
}
\arguments{
\item{df}{Original dataframe}

\item{col_cat}{Categorical columns index in \code{df}}
}
\value{
Dataframe where the categorical columns are the first columns
}
\description{
Change the position of categorical columns to adapt to the em.mix input form.
}
