% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mod.R
\name{prob_vector_cat}
\alias{prob_vector_cat}
\title{prob_vector_cat}
\usage{
prob_vector_cat(obs, tensor, dict_name_cat)
}
\arguments{
\item{obs}{One observation of all categorical variables.}

\item{tensor}{A tensor (multidimensional array) of probability, with dimension \code{nlevels(Y1) x nlevels(Y2) x...x nlevels(Yk)} if there are k categorical columns.}

\item{dict_name_cat}{A dictionary of categorical variable names and their corresponding onehot name. For example, \code{Y7: Y7_1, Y7_2,...}.}
}
\value{
Probability vector for all categorical columns.
}
\description{
Extract the probability vector from a tensor of probability. For example,
if there are only 2 categorical variables \code{(Y1, Y2)} with \code{nlevels(Y1)=2, nlevels(Y2)=3},
 then the tensor is a matrix of probability. \code{tensor[i,j]} is the probability of \code{Y1=i, Y2=j}.
 If \code{obs=(NA,2)}, then \code{prob_vector_cat} returns the probability vector
 of \code{Y1} when \code{Y2=2} and the onehot probability vector of \code{Y2=2}.
 If \code{obs=(NA,NA)}, then \code{prob_vector_cat} returns the marginal probability vector
 of \code{Y1} and that of \code{Y2=2}.
}
