% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_related.R
\name{result_mice}
\alias{result_mice}
\title{result_mice: Multiple imputation with bayesian method}
\usage{
result_mice(res, impnum, col_cat = c())
}
\arguments{
\item{res}{Result returned by \code{mice} function}

\item{impnum}{Number of multiple imputations}

\item{col_cat}{Categorical columns index}
}
\value{
\code{ximp} Final imputed data matrix, which is obtained with Robin's Rule.

\code{ximp.disj} Final imputed data matrix of same type as 'ximp' for the numeric columns.
 For the categorical columns, the prediction of probability for each category is shown in form of onehot probability vector.
}
\description{
\code{result_mice} is a function that retrieve the final imputed dataset after
running function \code{mice} in the 'mice' package.
As for categorical columns, both factor form and onehot probability vector form are returned

More details about the MICE implementation could be found in the documentation of function \code{mice} from
 'mice' package.
}
\references{
Van Buuren, S., Groothuis-Oudshoorn, K. (2011). \code{mice}:
Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of
Statistical Software}, \bold{45}(3), 1-67.
\url{https://www.jstatsoft.org/v45/i03/}

Van Buuren, S. (2018).
\href{https://stefvanbuuren.name/fimd/sec-FCS.html#sec:MICE}{\emph{Flexible Imputation of Missing Data. Second Edition.}}
Chapman & Hall/CRC. Boca Raton, FL.

Van Buuren, S., Brand, J.P.L., Groothuis-Oudshoorn C.G.M., Rubin, D.B. (2006)
Fully conditional specification in multivariate imputation.  \emph{Journal of
Statistical Computation and Simulation}, \bold{76}, 12, 1049--1064.

Van Buuren, S. (2007) Multiple imputation of discrete and continuous data by
fully conditional specification.  \emph{Statistical Methods in Medical
Research}, \bold{16}, 3, 219--242.

Van Buuren, S., Boshuizen, H.C., Knook, D.L. (1999) Multiple imputation of
missing blood pressure covariates in survival analysis.  \emph{Statistics in
Medicine}, \bold{18}, 681--694.

Brand, J.P.L. (1999) \emph{Development, implementation and evaluation of
multiple imputation strategies for the statistical analysis of incomplete
data sets.} Dissertation. Rotterdam: Erasmus University.

Statistical Analysis with Missing Data, by Little and Rubin, 2002
}
