% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_abalone.R
\docType{data}
\name{test_abalone}
\alias{test_abalone}
\title{Part of Abalone dataset for test}
\format{
A data frame with one categorical variable ('V1'), one interger variable ('V8') and 8 numerical variables.
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/abalone}{UCI Machine Learninng Repository}
}
\usage{
data('test_abalone')
}
\description{
First 100 lines of Abalone dataset from the UCI Machine Learning Repository.
}
\references{
Warwick J Nash, Tracy L Sellers, Simon R Talbot, Andrew J Cawthorn and Wes B Ford (1994)
"The Population Biology of Abalone (_Haliotis_ species) in Tasmania. I. Blacklip Abalone (_H. rubra_) from the North Coast and Islands of Bass Strait",
Fisheries Division, Technical Report No. 48 (ISSN 1034-3288)
}
\keyword{test_abalone}
