% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VA_fact}
\alias{VA_fact}
\title{VA_fact}
\usage{
VA_fact(x)
}
\arguments{
\item{x}{A vector of categorical variable}
}
\value{
Wilcox's VarNC
}
\description{
Calculate the Wilcox's VarNC for a vector of categorical values
}
