% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which.max.random}
\alias{which.max.random}
\title{which.max.random}
\usage{
which.max.random(vec)
}
\arguments{
\item{vec}{Numeric (logical, integer or double) vector.}
}
\value{
Index of the maximum value.
}
\description{
Determines the location, i.e., index of the maximum of a numeric (or logical) vector.
In case of tie, the index is selected randomly.
}
