% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissImp.R
\name{MissImp}
\alias{MissImp}
\title{Imputation of Missing Values with variance calculation}
\usage{
MissImp(
  df,
  imp_method = "missRanger",
  resample_method = "bootstrap",
  n_resample = 2 * round(log(nrow(df))),
  col_cat = c(),
  col_dis = c(),
  maxiter_tree = 10,
  maxiter_pca = 100,
  maxiter_mice = 10,
  ncp_pca = ncol(df)/2,
  learn_ncp = FALSE,
  cat_combine_by = "factor",
  var_cat = "wilcox_va",
  df_complete = NULL,
  num_mi = NULL
)
}
\arguments{
\item{df}{A \code{data.frame} with missing values to impute.}

\item{imp_method}{Imputation method to be chosen. The choices are: "missRanger", "kNN", "missForest", "PCA", "EM", "MI_EM", "MI_PCA", "MICE", "MI_Ranger" and "MI_Ranger_bis".}

\item{resample_method}{Resampling method to be chosen. The choices are: "bootstrap", "jackknife" and "none".}

\item{n_resample}{Number of datasets created by resampling.}

\item{col_cat}{Index of categorical columns.}

\item{col_dis}{Index of discrete columns with integer value.}

\item{maxiter_tree}{Max number of iterations with tree-based methods ("missRanger", "missForest", "MI_Ranger", "MI_Ranger_bis").}

\item{maxiter_pca}{Max number of iterations with PCA-based methods ("PCA", "MI_PCA").}

\item{maxiter_mice}{Max number of iterations with MICE imputation method.}

\item{ncp_pca}{Number of component for PCA-based methods ("PCA", "MI_PCA").}

\item{learn_ncp}{If \code{ncp_pca} is learned. This could lead to long execution time.}

\item{cat_combine_by}{Combine method for the categorical part of imputed resampled datasets.
'factor' means that the mode value of several predictions is taken as the final result. And 'onehot' means that we take the average of the probability vectors before
choosing the category that maximise the probability as the final result.}

\item{var_cat}{Method used to caculate the 'variance' on a prediction of categorical variable. 'wilcox_va' means the Wilcox's VarNC index and 'unalike' means unalikeability.}

\item{df_complete}{Complete dataset without missing values (if known). With this dataset, the performance of each imputation method could be estimated.}

\item{num_mi}{Number of multiple imputation.}
}
\value{
\code{df_result} The final imputation result.
}
\description{
\code{MissImp} is a function that could impute the mix-type missing values with a given imputation method
(signle imputation or multiple imputation). Bootstrap or Jackknife resampling method could also be applied. (To be completed)
}
